WIP: Bug fixes

+++ src/CAupdate.cpp
@@ -501,11 +501,11 @@ void CellCapture(int, int np, int, int, int, int nx, int MyYSlices, InterfacialR
                                        MyOrientation, GrainUnitVector, CritDiagonalLength);
                 if (np > 1) {
 
-                    double GhostGID = static_cast<double>(MyGrainID);
-                    double GhostDOCX = static_cast<double>(GlobalX + 0.5);
-                    double GhostDOCY = static_cast<double>(GlobalY + 0.5);
-                    double GhostDOCZ = static_cast<double>(GlobalZ + 0.5);
-                    double GhostDL = 0.01;
+                    int GhostGID = MyGrainID;
+                    float GhostDOCX = GlobalX + 0.5;
+                    float GhostDOCY = GlobalY + 0.5;
+                    float GhostDOCZ = GlobalZ + 0.5;
+                    float GhostDL = 0.01;
                     // Collect data for the ghost nodes, if necessary
                     loadghostnodes(GhostGID, GhostDOCX, GhostDOCY, GhostDOCZ, GhostDL, BufSizeX, MyYSlices, GlobalX,
                                    RankY, RankZ, AtNorthBoundary, AtSouthBoundary, BufferNorthSend_Octahedron,
